+++ src/components/TabIcon.vue
@@ -1,18 +1,21 @@
 <template>
   <div :class="bem( 'tab-icon', { 'is-loading': tab.status === 'loading' } )">
-    <img v-if="error_url !== icon_url" class="tab-icon__img" :src="icon_url" :style="{ height: `${ size }px`, width: `${ size }px` }" @error="onIconLoadError"/>
-    <svg v-if="tab.muted" class="tab-icon__state audio-mute-icon" xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16"
+    <img v-if="! icon.is_svg && error_url !== icon.url" class="tab-icon__img" :src="icon.url" :style="{ height: `${ size }px`, width: `${ size }px` }" @error="onIconLoadError"/>
+    <svg v-if="icon.is_svg" :class="bem( 'tab-icon__img', { 'is-svg': true, theme } )">
+      <use :xlink:href="`${ icon.url }#icon`"></use>
+    </svg>
+    <svg v-if="tab.muted" :class="[ bem( 'tab-icon__state', { theme } ), 'audio-mute-icon']" xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16"
         @click.prevent="unmuteTab( $event, tab.id )"
     >
-      <g :fill="context_fill">
+      <g>
         <path d="M13 8a2.813 2.813 0 0 0-.465-1.535l-.744.744A1.785 1.785 0 0 1 12 8a2.008 2.008 0 0 1-1.4 1.848.5.5 0 0 0 .343.939A3 3 0 0 0 13 8z"></path>
         <path d="M13.273 5.727A3.934 3.934 0 0 1 14 8a3.984 3.984 0 0 1-2.742 3.775.5.5 0 0 0 .316.949A4.985 4.985 0 0 0 15 8a4.93 4.93 0 0 0-1.012-2.988zm-4.603 7.99a.2.2 0 0 0 .33-.152V10l-2.154 2.154zm6.037-12.424a1 1 0 0 0-1.414 0L9 5.586V2.544a.25.25 0 0 0-.413-.19L5.5 5H4.191A2.191 2.191 0 0 0 2 7.191v1.618a2.186 2.186 0 0 0 1.659 2.118l-2.366 2.366a1 1 0 1 0 1.414 1.414l12-12a1 1 0 0 0 0-1.414z"></path>
       </g>
     </svg>
-    <svg v-else-if="tab.audible" class="tab-icon__state audio-icon" xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16"
+    <svg v-else-if="tab.audible" :class="[ bem( 'tab-icon__state', { theme } ), 'audio-icon']" xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16"
         @click.prevent="muteTab( $event, tab.id )"
     >
-      <g :fill="context_fill">
+      <g>
         <path d="M8.587 2.354L5.5 5H4.191A2.191 2.191 0 0 0 2 7.191v1.618A2.191 2.191 0 0 0 4.191 11H5.5l3.17 2.717a.2.2 0 0 0 .33-.152V2.544a.25.25 0 0 0-.413-.19zm2.988.921a.5.5 0 0 0-.316.949 3.97 3.97 0 0 1 0 7.551.5.5 0 0 0 .316.949 4.971 4.971 0 0 0 0-9.449z"></path>
         <path d="M13 8a3 3 0 0 0-2.056-2.787.5.5 0 1 0-.343.939A2.008 2.008 0 0 1 12 8a2.008 2.008 0 0 1-1.4 1.848.5.5 0 0 0 .343.939A3 3 0 0 0 13 8z"></path>
       </g>
@@ -36,9 +39,6 @@
 
 <script>
 import {
-  INK_90,
-} from "./photon-colors"
-import {
   bem,
 } from "./helpers.mjs"
 
@@ -48,55 +48,42 @@ export default {
   data() {
     return {
       error_url: null,
-      window_id: window.current_window_id,
     }
   },
   computed: {
-    context_fill() {
-      switch( this.theme ) {
-        // @todo handle system theme
-        case "dark":
-          return "rgb(255, 255, 255)"
-        default:
-          return INK_90
-      }
-    },
-    icon_url() {
+    icon() {
+      // Handle icons blocked by app security
       switch( this.tab.icon_url ) {
         case "chrome://mozapps/skin/extensions/extensionGeneric-16.svg":
-          return "/favicons/extensionGeneric.svg"
+          return { is_svg: true, url: "/favicons/firefox-addon.svg" }
         case "chrome://branding/content/icon32.png":
-          // @todo handle system theme
-          if( this.theme === "dark" ) {
-            return `/favicons/firefox-logo-glyph.svg`
-          }
-      }
-
-      // Add mapping for icons blocked by tracking protection
-      // Twitter
-      if( this.tab.url.startsWith( "https://twitter.com" ) ) {
-        return `/favicons/twitter.svg`
-      }
-      // Google
-      if( this.tab.url.startsWith( "https://mail.google.com" ) ) {
-        return `/favicons/google/mail.ico`
-      }
-      if( this.tab.url.startsWith( "https://docs.google.com/document" ) ) {
-        return `/favicons/google/document.ico`
-      }
-      if( this.tab.url.startsWith( "https://docs.google.com/presentation" ) ) {
-        return `/favicons/google/presentation.ico`
+          return { is_svg: true, url: "/favicons/firefox-logo.svg" }
       }
 
       if( this.tab.url.startsWith( "about:" ) ) {
         if( this.tab.url === 'about:debugging' || this.tab.url === 'about:config' || this.tab.url === 'about:newtab' ) {
-          // @todo handle system theme
-          if( this.theme === "dark" ) {
-            return `/favicons/firefox-logo-glyph.svg`
-          }
+          return { is_svg: true, url: "/favicons/firefox-logo.svg" }
         }
       }
-      return this.tab.icon_url
+
+      // Handle icons blocked by tracking protection
+      // Twitter
+      if( this.tab.url.startsWith( "https://twitter.com" ) ) {
+        return { is_svg: false, url: "/favicons/twitter.svg" }
+      }
+      // Google
+      if( this.tab.url.startsWith( "https://mail.google.com" ) ) {
+        return { is_svg: false, url: "/favicons/google/mail.ico" }
+      }
+      if( this.tab.url.startsWith( "https://docs.google.com/document" ) ) {
+        return { is_svg: false, url: "/favicons/google/document.ico" }
+      }
+      if( this.tab.url.startsWith( "https://docs.google.com/presentation" ) ) {
+        return { is_svg: false, url: "/favicons/google/presentation.ico" }
+      }
+
+      // Remaining icons
+      return { is_svg: false, url: this.tab.icon_url }
     },
   },
   created() {
@@ -106,12 +93,12 @@ export default {
     muteTab( event ) {
       console.info('muteTab', this.tab.id)
       event.stopPropagation()
-      window.background.muteTab( window.store, this.window_id, this.tab.id )
+      window.background.muteTab( window.store, window.current_window_id, this.tab.id )
     },
     unmuteTab( event ) {
       console.info('unmuteTab', this.tab.id)
       event.stopPropagation()
-      window.background.unmuteTab( window.store, this.window_id, this.tab.id )
+      window.background.unmuteTab( window.store, window.current_window_id, this.tab.id )
     },
     onIconLoadError( event ) {
       console.info('onIconLoadError', event, this.tab.id, this.tab.icon_url)
@@ -144,6 +131,26 @@ $tab-icon__state--size: 12px !default;
   &__img {
     width: $tab-icon--size;
     height: $tab-icon--size;
+
+    &--is-svg#{&}--theme-dark {
+      fill: $grey-30;
+    }
+
+    &--is-svg#{&}--theme-light {
+      fill: $ink-90;
+    }
+
+    @media (prefers-color-scheme: dark) {
+      &--is-svg#{&}--theme-system {
+        fill: $grey-30;
+      }
+    }
+
+    @media (prefers-color-scheme: light) {
+      &--is-svg#{&}--theme-system {
+        fill: $ink-90;
+      }
+    }
   }
 
   &__state {
@@ -152,6 +159,26 @@ $tab-icon__state--size: 12px !default;
     right: -8px;
     height: $tab-icon__state--size;
     width: $tab-icon__state--size;
+
+    &--theme-dark {
+      fill: rgb(255, 255, 255);
+    }
+
+    &--theme-light {
+      fill: $ink-90;
+    }
+
+    @media (prefers-color-scheme: dark) {
+      &--theme-system {
+        fill: rgb(255, 255, 255);
+      }
+    }
+
+    @media (prefers-color-scheme: light) {
+      &--theme-system {
+        fill: $ink-90;
+      }
+    }
   }
 
   &__spinner {